Generate latest types from archive
