+++ src/types/asset-tx-payment/events.ts
@@ -3,6 +3,7 @@ import * as v47 from '../v47'
 import * as v48 from '../v48'
 import * as v54 from '../v54'
 import * as v56 from '../v56'
+import * as v60 from '../v60'
 
 export const assetTxFeePaid =  {
     name: 'AssetTxPayment.AssetTxFeePaid',
@@ -56,6 +57,19 @@ export const assetTxFeePaid =  {
             actualFee: sts.bigint(),
             tip: sts.bigint(),
             assetId: sts.option(() => sts.number()),
+        })
+    ),
+    /**
+     * A transaction fee `actual_fee`, of which `tip` was added to the minimum inclusion fee,
+     * has been paid by `who` in an asset `asset_id`.
+     */
+    v60: new EventType(
+        'AssetTxPayment.AssetTxFeePaid',
+        sts.struct({
+            who: v60.AccountId32,
+            actualFee: sts.bigint(),
+            tip: sts.bigint(),
+            assetId: sts.option(() => v60.V3MultiLocation),
         })
     ),
 }
