+++ src/types/index.ts
@@ -1,4 +1,4 @@
-export * as v23 from './v23'
+export * as v26 from './v26'
 export * as v34 from './v34'
 export * as v42 from './v42'
 export * as v46 from './v46'
@@ -10,11 +10,11 @@ export * as v33 from './v33'
 export * as v41 from './v41'
 export * as v51 from './v51'
 export * as v54 from './v54'
+export * as v60 from './v60'
 export * as v37 from './v37'
 export * as v39 from './v39'
 export * as v50 from './v50'
 export * as v53 from './v53'
-export * as v26 from './v26'
 export * as v49 from './v49'
 export * as v29 from './v29'
 export * as v38 from './v38'
