+++ src/types/parimutuel/events.ts
@@ -2,6 +2,7 @@ import {sts, Block, Bytes, Option, Result, EventType, RuntimeCtx} from '../suppo
 import * as v51 from '../v51'
 import * as v54 from '../v54'
 import * as v56 from '../v56'
+import * as v60 from '../v60'
 
 export const outcomeBought =  {
     name: 'Parimutuel.OutcomeBought',
@@ -40,6 +41,19 @@ export const outcomeBought =  {
             marketId: sts.bigint(),
             buyer: v56.AccountId32,
             asset: v56.Asset,
+            amountMinusFees: sts.bigint(),
+            fees: sts.bigint(),
+        })
+    ),
+    /**
+     * An outcome was bought.
+     */
+    v60: new EventType(
+        'Parimutuel.OutcomeBought',
+        sts.struct({
+            marketId: sts.bigint(),
+            buyer: v60.AccountId32,
+            asset: v60.Asset,
             amountMinusFees: sts.bigint(),
             fees: sts.bigint(),
         })
