+++ src/types/swaps/calls.ts
@@ -1,5 +1,5 @@
 import {sts, Block, Bytes, Option, Result, CallType, RuntimeCtx} from '../support'
-import * as v23 from '../v23'
+import * as v26 from '../v26'
 
 export const poolExit =  {
     name: 'Swaps.pool_exit',
@@ -17,12 +17,12 @@ export const poolExit =  {
      *  * `min_assets_out`: List of asset lower bounds. No asset should be lower than the
      *  provided values.
      */
-    v23: new CallType(
+    v26: new CallType(
         'Swaps.pool_exit',
         sts.struct({
-            poolId: v23.PoolId,
-            poolAmount: v23.BalanceOf,
-            minAssetsOut: sts.array(() => v23.BalanceOf),
+            poolId: v26.PoolId,
+            poolAmount: v26.BalanceOf,
+            minAssetsOut: sts.array(() => v26.BalanceOf),
         })
     ),
 }
