+++ src/types/v60.ts
@@ -1,0 +1,1201 @@
+import {sts, Result, Option, Bytes, BitSequence} from './support'
+
+export type Asset = Asset_CategoricalOutcome | Asset_CombinatorialOutcomeLegacy | Asset_CombinatorialToken | Asset_ForeignAsset | Asset_ParimutuelShare | Asset_PoolShare | Asset_ScalarOutcome | Asset_Ztg
+
+export interface Asset_CategoricalOutcome {
+    __kind: 'CategoricalOutcome'
+    value: [bigint, number]
+}
+
+export interface Asset_CombinatorialOutcomeLegacy {
+    __kind: 'CombinatorialOutcomeLegacy'
+}
+
+export interface Asset_CombinatorialToken {
+    __kind: 'CombinatorialToken'
+    value: Bytes
+}
+
+export interface Asset_ForeignAsset {
+    __kind: 'ForeignAsset'
+    value: number
+}
+
+export interface Asset_ParimutuelShare {
+    __kind: 'ParimutuelShare'
+    value: [bigint, number]
+}
+
+export interface Asset_PoolShare {
+    __kind: 'PoolShare'
+    value: bigint
+}
+
+export interface Asset_ScalarOutcome {
+    __kind: 'ScalarOutcome'
+    value: [bigint, ScalarPosition]
+}
+
+export interface Asset_Ztg {
+    __kind: 'Ztg'
+}
+
+export type ScalarPosition = ScalarPosition_Long | ScalarPosition_Short
+
+export interface ScalarPosition_Long {
+    __kind: 'Long'
+}
+
+export interface ScalarPosition_Short {
+    __kind: 'Short'
+}
+
+export interface AssetMetadata {
+    decimals: number
+    name: Bytes
+    symbol: Bytes
+    existentialDeposit: bigint
+    location?: (VersionedMultiLocation | undefined)
+    additional: CustomMetadata
+}
+
+export interface CustomMetadata {
+    xcm: XcmMetadata
+    allowAsBaseAsset: boolean
+}
+
+export interface XcmMetadata {
+    feeFactor?: (bigint | undefined)
+}
+
+export type VersionedMultiLocation = VersionedMultiLocation_V2 | VersionedMultiLocation_V3
+
+export interface VersionedMultiLocation_V2 {
+    __kind: 'V2'
+    value: V2MultiLocation
+}
+
+export interface VersionedMultiLocation_V3 {
+    __kind: 'V3'
+    value: V3MultiLocation
+}
+
+export interface V3MultiLocation {
+    parents: number
+    interior: V3Junctions
+}
+
+export type V3Junctions = V3Junctions_Here | V3Junctions_X1 | V3Junctions_X2 | V3Junctions_X3 | V3Junctions_X4 | V3Junctions_X5 | V3Junctions_X6 | V3Junctions_X7 | V3Junctions_X8
+
+export interface V3Junctions_Here {
+    __kind: 'Here'
+}
+
+export interface V3Junctions_X1 {
+    __kind: 'X1'
+    value: V3Junction
+}
+
+export interface V3Junctions_X2 {
+    __kind: 'X2'
+    value: [V3Junction, V3Junction]
+}
+
+export interface V3Junctions_X3 {
+    __kind: 'X3'
+    value: [V3Junction, V3Junction, V3Junction]
+}
+
+export interface V3Junctions_X4 {
+    __kind: 'X4'
+    value: [V3Junction, V3Junction, V3Junction, V3Junction]
+}
+
+export interface V3Junctions_X5 {
+    __kind: 'X5'
+    value: [V3Junction, V3Junction, V3Junction, V3Junction, V3Junction]
+}
+
+export interface V3Junctions_X6 {
+    __kind: 'X6'
+    value: [V3Junction, V3Junction, V3Junction, V3Junction, V3Junction, V3Junction]
+}
+
+export interface V3Junctions_X7 {
+    __kind: 'X7'
+    value: [V3Junction, V3Junction, V3Junction, V3Junction, V3Junction, V3Junction, V3Junction]
+}
+
+export interface V3Junctions_X8 {
+    __kind: 'X8'
+    value: [V3Junction, V3Junction, V3Junction, V3Junction, V3Junction, V3Junction, V3Junction, V3Junction]
+}
+
+export type V3Junction = V3Junction_AccountId32 | V3Junction_AccountIndex64 | V3Junction_AccountKey20 | V3Junction_GeneralIndex | V3Junction_GeneralKey | V3Junction_GlobalConsensus | V3Junction_OnlyChild | V3Junction_PalletInstance | V3Junction_Parachain | V3Junction_Plurality
+
+export interface V3Junction_AccountId32 {
+    __kind: 'AccountId32'
+    network?: (V3NetworkId | undefined)
+    id: Bytes
+}
+
+export interface V3Junction_AccountIndex64 {
+    __kind: 'AccountIndex64'
+    network?: (V3NetworkId | undefined)
+    index: bigint
+}
+
+export interface V3Junction_AccountKey20 {
+    __kind: 'AccountKey20'
+    network?: (V3NetworkId | undefined)
+    key: Bytes
+}
+
+export interface V3Junction_GeneralIndex {
+    __kind: 'GeneralIndex'
+    value: bigint
+}
+
+export interface V3Junction_GeneralKey {
+    __kind: 'GeneralKey'
+    length: number
+    data: Bytes
+}
+
+export interface V3Junction_GlobalConsensus {
+    __kind: 'GlobalConsensus'
+    value: V3NetworkId
+}
+
+export interface V3Junction_OnlyChild {
+    __kind: 'OnlyChild'
+}
+
+export interface V3Junction_PalletInstance {
+    __kind: 'PalletInstance'
+    value: number
+}
+
+export interface V3Junction_Parachain {
+    __kind: 'Parachain'
+    value: number
+}
+
+export interface V3Junction_Plurality {
+    __kind: 'Plurality'
+    id: V3BodyId
+    part: V3BodyPart
+}
+
+export type V3BodyPart = V3BodyPart_AtLeastProportion | V3BodyPart_Fraction | V3BodyPart_Members | V3BodyPart_MoreThanProportion | V3BodyPart_Voice
+
+export interface V3BodyPart_AtLeastProportion {
+    __kind: 'AtLeastProportion'
+    nom: number
+    denom: number
+}
+
+export interface V3BodyPart_Fraction {
+    __kind: 'Fraction'
+    nom: number
+    denom: number
+}
+
+export interface V3BodyPart_Members {
+    __kind: 'Members'
+    count: number
+}
+
+export interface V3BodyPart_MoreThanProportion {
+    __kind: 'MoreThanProportion'
+    nom: number
+    denom: number
+}
+
+export interface V3BodyPart_Voice {
+    __kind: 'Voice'
+}
+
+export type V3BodyId = V3BodyId_Administration | V3BodyId_Defense | V3BodyId_Executive | V3BodyId_Index | V3BodyId_Judicial | V3BodyId_Legislative | V3BodyId_Moniker | V3BodyId_Technical | V3BodyId_Treasury | V3BodyId_Unit
+
+export interface V3BodyId_Administration {
+    __kind: 'Administration'
+}
+
+export interface V3BodyId_Defense {
+    __kind: 'Defense'
+}
+
+export interface V3BodyId_Executive {
+    __kind: 'Executive'
+}
+
+export interface V3BodyId_Index {
+    __kind: 'Index'
+    value: number
+}
+
+export interface V3BodyId_Judicial {
+    __kind: 'Judicial'
+}
+
+export interface V3BodyId_Legislative {
+    __kind: 'Legislative'
+}
+
+export interface V3BodyId_Moniker {
+    __kind: 'Moniker'
+    value: Bytes
+}
+
+export interface V3BodyId_Technical {
+    __kind: 'Technical'
+}
+
+export interface V3BodyId_Treasury {
+    __kind: 'Treasury'
+}
+
+export interface V3BodyId_Unit {
+    __kind: 'Unit'
+}
+
+export type V3NetworkId = V3NetworkId_BitcoinCash | V3NetworkId_BitcoinCore | V3NetworkId_ByFork | V3NetworkId_ByGenesis | V3NetworkId_Ethereum | V3NetworkId_Kusama | V3NetworkId_Polkadot | V3NetworkId_Rococo | V3NetworkId_Westend | V3NetworkId_Wococo
+
+export interface V3NetworkId_BitcoinCash {
+    __kind: 'BitcoinCash'
+}
+
+export interface V3NetworkId_BitcoinCore {
+    __kind: 'BitcoinCore'
+}
+
+export interface V3NetworkId_ByFork {
+    __kind: 'ByFork'
+    blockNumber: bigint
+    blockHash: Bytes
+}
+
+export interface V3NetworkId_ByGenesis {
+    __kind: 'ByGenesis'
+    value: Bytes
+}
+
+export interface V3NetworkId_Ethereum {
+    __kind: 'Ethereum'
+    chainId: bigint
+}
+
+export interface V3NetworkId_Kusama {
+    __kind: 'Kusama'
+}
+
+export interface V3NetworkId_Polkadot {
+    __kind: 'Polkadot'
+}
+
+export interface V3NetworkId_Rococo {
+    __kind: 'Rococo'
+}
+
+export interface V3NetworkId_Westend {
+    __kind: 'Westend'
+}
+
+export interface V3NetworkId_Wococo {
+    __kind: 'Wococo'
+}
+
+export interface V2MultiLocation {
+    parents: number
+    interior: V2Junctions
+}
+
+export type V2Junctions = V2Junctions_Here | V2Junctions_X1 | V2Junctions_X2 | V2Junctions_X3 | V2Junctions_X4 | V2Junctions_X5 | V2Junctions_X6 | V2Junctions_X7 | V2Junctions_X8
+
+export interface V2Junctions_Here {
+    __kind: 'Here'
+}
+
+export interface V2Junctions_X1 {
+    __kind: 'X1'
+    value: V2Junction
+}
+
+export interface V2Junctions_X2 {
+    __kind: 'X2'
+    value: [V2Junction, V2Junction]
+}
+
+export interface V2Junctions_X3 {
+    __kind: 'X3'
+    value: [V2Junction, V2Junction, V2Junction]
+}
+
+export interface V2Junctions_X4 {
+    __kind: 'X4'
+    value: [V2Junction, V2Junction, V2Junction, V2Junction]
+}
+
+export interface V2Junctions_X5 {
+    __kind: 'X5'
+    value: [V2Junction, V2Junction, V2Junction, V2Junction, V2Junction]
+}
+
+export interface V2Junctions_X6 {
+    __kind: 'X6'
+    value: [V2Junction, V2Junction, V2Junction, V2Junction, V2Junction, V2Junction]
+}
+
+export interface V2Junctions_X7 {
+    __kind: 'X7'
+    value: [V2Junction, V2Junction, V2Junction, V2Junction, V2Junction, V2Junction, V2Junction]
+}
+
+export interface V2Junctions_X8 {
+    __kind: 'X8'
+    value: [V2Junction, V2Junction, V2Junction, V2Junction, V2Junction, V2Junction, V2Junction, V2Junction]
+}
+
+export type V2Junction = V2Junction_AccountId32 | V2Junction_AccountIndex64 | V2Junction_AccountKey20 | V2Junction_GeneralIndex | V2Junction_GeneralKey | V2Junction_OnlyChild | V2Junction_PalletInstance | V2Junction_Parachain | V2Junction_Plurality
+
+export interface V2Junction_AccountId32 {
+    __kind: 'AccountId32'
+    network: V2NetworkId
+    id: Bytes
+}
+
+export interface V2Junction_AccountIndex64 {
+    __kind: 'AccountIndex64'
+    network: V2NetworkId
+    index: bigint
+}
+
+export interface V2Junction_AccountKey20 {
+    __kind: 'AccountKey20'
+    network: V2NetworkId
+    key: Bytes
+}
+
+export interface V2Junction_GeneralIndex {
+    __kind: 'GeneralIndex'
+    value: bigint
+}
+
+export interface V2Junction_GeneralKey {
+    __kind: 'GeneralKey'
+    value: WeakBoundedVec
+}
+
+export interface V2Junction_OnlyChild {
+    __kind: 'OnlyChild'
+}
+
+export interface V2Junction_PalletInstance {
+    __kind: 'PalletInstance'
+    value: number
+}
+
+export interface V2Junction_Parachain {
+    __kind: 'Parachain'
+    value: number
+}
+
+export interface V2Junction_Plurality {
+    __kind: 'Plurality'
+    id: V2BodyId
+    part: V2BodyPart
+}
+
+export type V2BodyPart = V2BodyPart_AtLeastProportion | V2BodyPart_Fraction | V2BodyPart_Members | V2BodyPart_MoreThanProportion | V2BodyPart_Voice
+
+export interface V2BodyPart_AtLeastProportion {
+    __kind: 'AtLeastProportion'
+    nom: number
+    denom: number
+}
+
+export interface V2BodyPart_Fraction {
+    __kind: 'Fraction'
+    nom: number
+    denom: number
+}
+
+export interface V2BodyPart_Members {
+    __kind: 'Members'
+    count: number
+}
+
+export interface V2BodyPart_MoreThanProportion {
+    __kind: 'MoreThanProportion'
+    nom: number
+    denom: number
+}
+
+export interface V2BodyPart_Voice {
+    __kind: 'Voice'
+}
+
+export type V2BodyId = V2BodyId_Administration | V2BodyId_Defense | V2BodyId_Executive | V2BodyId_Index | V2BodyId_Judicial | V2BodyId_Legislative | V2BodyId_Named | V2BodyId_Technical | V2BodyId_Treasury | V2BodyId_Unit
+
+export interface V2BodyId_Administration {
+    __kind: 'Administration'
+}
+
+export interface V2BodyId_Defense {
+    __kind: 'Defense'
+}
+
+export interface V2BodyId_Executive {
+    __kind: 'Executive'
+}
+
+export interface V2BodyId_Index {
+    __kind: 'Index'
+    value: number
+}
+
+export interface V2BodyId_Judicial {
+    __kind: 'Judicial'
+}
+
+export interface V2BodyId_Legislative {
+    __kind: 'Legislative'
+}
+
+export interface V2BodyId_Named {
+    __kind: 'Named'
+    value: WeakBoundedVec
+}
+
+export interface V2BodyId_Technical {
+    __kind: 'Technical'
+}
+
+export interface V2BodyId_Treasury {
+    __kind: 'Treasury'
+}
+
+export interface V2BodyId_Unit {
+    __kind: 'Unit'
+}
+
+export type WeakBoundedVec = Bytes
+
+export type V2NetworkId = V2NetworkId_Any | V2NetworkId_Kusama | V2NetworkId_Named | V2NetworkId_Polkadot
+
+export interface V2NetworkId_Any {
+    __kind: 'Any'
+}
+
+export interface V2NetworkId_Kusama {
+    __kind: 'Kusama'
+}
+
+export interface V2NetworkId_Named {
+    __kind: 'Named'
+    value: WeakBoundedVec
+}
+
+export interface V2NetworkId_Polkadot {
+    __kind: 'Polkadot'
+}
+
+export const AssetMetadata: sts.Type<AssetMetadata> = sts.struct(() => {
+    return  {
+        decimals: sts.number(),
+        name: sts.bytes(),
+        symbol: sts.bytes(),
+        existentialDeposit: sts.bigint(),
+        location: sts.option(() => VersionedMultiLocation),
+        additional: CustomMetadata,
+    }
+})
+
+export const CustomMetadata: sts.Type<CustomMetadata> = sts.struct(() => {
+    return  {
+        xcm: XcmMetadata,
+        allowAsBaseAsset: sts.boolean(),
+    }
+})
+
+export const XcmMetadata: sts.Type<XcmMetadata> = sts.struct(() => {
+    return  {
+        feeFactor: sts.option(() => sts.bigint()),
+    }
+})
+
+export const VersionedMultiLocation: sts.Type<VersionedMultiLocation> = sts.closedEnum(() => {
+    return  {
+        V2: V2MultiLocation,
+        V3: V3MultiLocation,
+    }
+})
+
+export const V2MultiLocation: sts.Type<V2MultiLocation> = sts.struct(() => {
+    return  {
+        parents: sts.number(),
+        interior: V2Junctions,
+    }
+})
+
+export const V2Junctions: sts.Type<V2Junctions> = sts.closedEnum(() => {
+    return  {
+        Here: sts.unit(),
+        X1: V2Junction,
+        X2: sts.tuple(() => [V2Junction, V2Junction]),
+        X3: sts.tuple(() => [V2Junction, V2Junction, V2Junction]),
+        X4: sts.tuple(() => [V2Junction, V2Junction, V2Junction, V2Junction]),
+        X5: sts.tuple(() => [V2Junction, V2Junction, V2Junction, V2Junction, V2Junction]),
+        X6: sts.tuple(() => [V2Junction, V2Junction, V2Junction, V2Junction, V2Junction, V2Junction]),
+        X7: sts.tuple(() => [V2Junction, V2Junction, V2Junction, V2Junction, V2Junction, V2Junction, V2Junction]),
+        X8: sts.tuple(() => [V2Junction, V2Junction, V2Junction, V2Junction, V2Junction, V2Junction, V2Junction, V2Junction]),
+    }
+})
+
+export const V2Junction: sts.Type<V2Junction> = sts.closedEnum(() => {
+    return  {
+        AccountId32: sts.enumStruct({
+            network: V2NetworkId,
+            id: sts.bytes(),
+        }),
+        AccountIndex64: sts.enumStruct({
+            network: V2NetworkId,
+            index: sts.bigint(),
+        }),
+        AccountKey20: sts.enumStruct({
+            network: V2NetworkId,
+            key: sts.bytes(),
+        }),
+        GeneralIndex: sts.bigint(),
+        GeneralKey: WeakBoundedVec,
+        OnlyChild: sts.unit(),
+        PalletInstance: sts.number(),
+        Parachain: sts.number(),
+        Plurality: sts.enumStruct({
+            id: V2BodyId,
+            part: V2BodyPart,
+        }),
+    }
+})
+
+export const V2BodyPart: sts.Type<V2BodyPart> = sts.closedEnum(() => {
+    return  {
+        AtLeastProportion: sts.enumStruct({
+            nom: sts.number(),
+            denom: sts.number(),
+        }),
+        Fraction: sts.enumStruct({
+            nom: sts.number(),
+            denom: sts.number(),
+        }),
+        Members: sts.enumStruct({
+            count: sts.number(),
+        }),
+        MoreThanProportion: sts.enumStruct({
+            nom: sts.number(),
+            denom: sts.number(),
+        }),
+        Voice: sts.unit(),
+    }
+})
+
+export const V2BodyId: sts.Type<V2BodyId> = sts.closedEnum(() => {
+    return  {
+        Administration: sts.unit(),
+        Defense: sts.unit(),
+        Executive: sts.unit(),
+        Index: sts.number(),
+        Judicial: sts.unit(),
+        Legislative: sts.unit(),
+        Named: WeakBoundedVec,
+        Technical: sts.unit(),
+        Treasury: sts.unit(),
+        Unit: sts.unit(),
+    }
+})
+
+export const WeakBoundedVec = sts.bytes()
+
+export const V2NetworkId: sts.Type<V2NetworkId> = sts.closedEnum(() => {
+    return  {
+        Any: sts.unit(),
+        Kusama: sts.unit(),
+        Named: WeakBoundedVec,
+        Polkadot: sts.unit(),
+    }
+})
+
+export interface Market {
+    marketId: bigint
+    baseAsset: Asset
+    creator: AccountId32
+    creation: MarketCreation
+    creatorFee: Perbill
+    oracle: AccountId32
+    metadata: Bytes
+    marketType: MarketType
+    period: MarketPeriod
+    deadlines: Deadlines
+    scoringRule: ScoringRule
+    status: MarketStatus
+    report?: (Report | undefined)
+    resolvedOutcome?: (OutcomeReport | undefined)
+    disputeMechanism?: (MarketDisputeMechanism | undefined)
+    bonds: MarketBonds
+    earlyClose?: (EarlyClose | undefined)
+}
+
+export interface EarlyClose {
+    old: MarketPeriod
+    new: MarketPeriod
+    state: EarlyCloseState
+}
+
+export type EarlyCloseState = EarlyCloseState_Disputed | EarlyCloseState_Rejected | EarlyCloseState_ScheduledAsMarketCreator | EarlyCloseState_ScheduledAsOther
+
+export interface EarlyCloseState_Disputed {
+    __kind: 'Disputed'
+}
+
+export interface EarlyCloseState_Rejected {
+    __kind: 'Rejected'
+}
+
+export interface EarlyCloseState_ScheduledAsMarketCreator {
+    __kind: 'ScheduledAsMarketCreator'
+}
+
+export interface EarlyCloseState_ScheduledAsOther {
+    __kind: 'ScheduledAsOther'
+}
+
+export interface MarketBonds {
+    creation?: (Bond | undefined)
+    oracle?: (Bond | undefined)
+    outsider?: (Bond | undefined)
+    dispute?: (Bond | undefined)
+    closeRequest?: (Bond | undefined)
+    closeDispute?: (Bond | undefined)
+}
+
+export interface Bond {
+    who: AccountId32
+    value: bigint
+    isSettled: boolean
+}
+
+export type MarketDisputeMechanism = MarketDisputeMechanism_Authorized | MarketDisputeMechanism_Court
+
+export interface MarketDisputeMechanism_Authorized {
+    __kind: 'Authorized'
+}
+
+export interface MarketDisputeMechanism_Court {
+    __kind: 'Court'
+}
+
+export type OutcomeReport = OutcomeReport_Categorical | OutcomeReport_Scalar
+
+export interface OutcomeReport_Categorical {
+    __kind: 'Categorical'
+    value: number
+}
+
+export interface OutcomeReport_Scalar {
+    __kind: 'Scalar'
+    value: bigint
+}
+
+export interface Report {
+    at: bigint
+    by: AccountId32
+    outcome: OutcomeReport
+}
+
+export type MarketStatus = MarketStatus_Active | MarketStatus_Closed | MarketStatus_Disputed | MarketStatus_Proposed | MarketStatus_Reported | MarketStatus_Resolved
+
+export interface MarketStatus_Active {
+    __kind: 'Active'
+}
+
+export interface MarketStatus_Closed {
+    __kind: 'Closed'
+}
+
+export interface MarketStatus_Disputed {
+    __kind: 'Disputed'
+}
+
+export interface MarketStatus_Proposed {
+    __kind: 'Proposed'
+}
+
+export interface MarketStatus_Reported {
+    __kind: 'Reported'
+}
+
+export interface MarketStatus_Resolved {
+    __kind: 'Resolved'
+}
+
+export type ScoringRule = ScoringRule_AmmCdaHybrid | ScoringRule_Parimutuel
+
+export interface ScoringRule_AmmCdaHybrid {
+    __kind: 'AmmCdaHybrid'
+}
+
+export interface ScoringRule_Parimutuel {
+    __kind: 'Parimutuel'
+}
+
+export interface Deadlines {
+    gracePeriod: bigint
+    oracleDuration: bigint
+    disputeDuration: bigint
+}
+
+export type MarketPeriod = MarketPeriod_Block | MarketPeriod_Timestamp
+
+export interface MarketPeriod_Block {
+    __kind: 'Block'
+    value: Range
+}
+
+export interface MarketPeriod_Timestamp {
+    __kind: 'Timestamp'
+    value: Range
+}
+
+export interface Range {
+    start: bigint
+    end: bigint
+}
+
+export type MarketType = MarketType_Categorical | MarketType_Scalar
+
+export interface MarketType_Categorical {
+    __kind: 'Categorical'
+    value: number
+}
+
+export interface MarketType_Scalar {
+    __kind: 'Scalar'
+    value: RangeInclusive
+}
+
+export interface RangeInclusive {
+    start: bigint
+    end: bigint
+}
+
+export type Perbill = number
+
+export type MarketCreation = MarketCreation_Advised | MarketCreation_Permissionless
+
+export interface MarketCreation_Advised {
+    __kind: 'Advised'
+}
+
+export interface MarketCreation_Permissionless {
+    __kind: 'Permissionless'
+}
+
+export type AccountId32 = Bytes
+
+export const Order: sts.Type<Order> = sts.struct(() => {
+    return  {
+        marketId: sts.bigint(),
+        maker: AccountId32,
+        makerAsset: Asset,
+        makerAmount: sts.bigint(),
+        takerAsset: Asset,
+        takerAmount: sts.bigint(),
+    }
+})
+
+export interface Order {
+    marketId: bigint
+    maker: AccountId32
+    makerAsset: Asset
+    makerAmount: bigint
+    takerAsset: Asset
+    takerAmount: bigint
+}
+
+export const V3MultiLocation: sts.Type<V3MultiLocation> = sts.struct(() => {
+    return  {
+        parents: sts.number(),
+        interior: V3Junctions,
+    }
+})
+
+export const V3Junctions: sts.Type<V3Junctions> = sts.closedEnum(() => {
+    return  {
+        Here: sts.unit(),
+        X1: V3Junction,
+        X2: sts.tuple(() => [V3Junction, V3Junction]),
+        X3: sts.tuple(() => [V3Junction, V3Junction, V3Junction]),
+        X4: sts.tuple(() => [V3Junction, V3Junction, V3Junction, V3Junction]),
+        X5: sts.tuple(() => [V3Junction, V3Junction, V3Junction, V3Junction, V3Junction]),
+        X6: sts.tuple(() => [V3Junction, V3Junction, V3Junction, V3Junction, V3Junction, V3Junction]),
+        X7: sts.tuple(() => [V3Junction, V3Junction, V3Junction, V3Junction, V3Junction, V3Junction, V3Junction]),
+        X8: sts.tuple(() => [V3Junction, V3Junction, V3Junction, V3Junction, V3Junction, V3Junction, V3Junction, V3Junction]),
+    }
+})
+
+export const V3Junction: sts.Type<V3Junction> = sts.closedEnum(() => {
+    return  {
+        AccountId32: sts.enumStruct({
+            network: sts.option(() => V3NetworkId),
+            id: sts.bytes(),
+        }),
+        AccountIndex64: sts.enumStruct({
+            network: sts.option(() => V3NetworkId),
+            index: sts.bigint(),
+        }),
+        AccountKey20: sts.enumStruct({
+            network: sts.option(() => V3NetworkId),
+            key: sts.bytes(),
+        }),
+        GeneralIndex: sts.bigint(),
+        GeneralKey: sts.enumStruct({
+            length: sts.number(),
+            data: sts.bytes(),
+        }),
+        GlobalConsensus: V3NetworkId,
+        OnlyChild: sts.unit(),
+        PalletInstance: sts.number(),
+        Parachain: sts.number(),
+        Plurality: sts.enumStruct({
+            id: V3BodyId,
+            part: V3BodyPart,
+        }),
+    }
+})
+
+export const V3BodyPart: sts.Type<V3BodyPart> = sts.closedEnum(() => {
+    return  {
+        AtLeastProportion: sts.enumStruct({
+            nom: sts.number(),
+            denom: sts.number(),
+        }),
+        Fraction: sts.enumStruct({
+            nom: sts.number(),
+            denom: sts.number(),
+        }),
+        Members: sts.enumStruct({
+            count: sts.number(),
+        }),
+        MoreThanProportion: sts.enumStruct({
+            nom: sts.number(),
+            denom: sts.number(),
+        }),
+        Voice: sts.unit(),
+    }
+})
+
+export const V3BodyId: sts.Type<V3BodyId> = sts.closedEnum(() => {
+    return  {
+        Administration: sts.unit(),
+        Defense: sts.unit(),
+        Executive: sts.unit(),
+        Index: sts.number(),
+        Judicial: sts.unit(),
+        Legislative: sts.unit(),
+        Moniker: sts.bytes(),
+        Technical: sts.unit(),
+        Treasury: sts.unit(),
+        Unit: sts.unit(),
+    }
+})
+
+export const V3NetworkId: sts.Type<V3NetworkId> = sts.closedEnum(() => {
+    return  {
+        BitcoinCash: sts.unit(),
+        BitcoinCore: sts.unit(),
+        ByFork: sts.enumStruct({
+            blockNumber: sts.bigint(),
+            blockHash: sts.bytes(),
+        }),
+        ByGenesis: sts.bytes(),
+        Ethereum: sts.enumStruct({
+            chainId: sts.bigint(),
+        }),
+        Kusama: sts.unit(),
+        Polkadot: sts.unit(),
+        Rococo: sts.unit(),
+        Westend: sts.unit(),
+        Wococo: sts.unit(),
+    }
+})
+
+export const Market: sts.Type<Market> = sts.struct(() => {
+    return  {
+        marketId: sts.bigint(),
+        baseAsset: Asset,
+        creator: AccountId32,
+        creation: MarketCreation,
+        creatorFee: Perbill,
+        oracle: AccountId32,
+        metadata: sts.bytes(),
+        marketType: MarketType,
+        period: MarketPeriod,
+        deadlines: Deadlines,
+        scoringRule: ScoringRule,
+        status: MarketStatus,
+        report: sts.option(() => Report),
+        resolvedOutcome: sts.option(() => OutcomeReport),
+        disputeMechanism: sts.option(() => MarketDisputeMechanism),
+        bonds: MarketBonds,
+        earlyClose: sts.option(() => EarlyClose),
+    }
+})
+
+export const EarlyClose: sts.Type<EarlyClose> = sts.struct(() => {
+    return  {
+        old: MarketPeriod,
+        new: MarketPeriod,
+        state: EarlyCloseState,
+    }
+})
+
+export const EarlyCloseState: sts.Type<EarlyCloseState> = sts.closedEnum(() => {
+    return  {
+        Disputed: sts.unit(),
+        Rejected: sts.unit(),
+        ScheduledAsMarketCreator: sts.unit(),
+        ScheduledAsOther: sts.unit(),
+    }
+})
+
+export const MarketBonds: sts.Type<MarketBonds> = sts.struct(() => {
+    return  {
+        creation: sts.option(() => Bond),
+        oracle: sts.option(() => Bond),
+        outsider: sts.option(() => Bond),
+        dispute: sts.option(() => Bond),
+        closeRequest: sts.option(() => Bond),
+        closeDispute: sts.option(() => Bond),
+    }
+})
+
+export const Bond: sts.Type<Bond> = sts.struct(() => {
+    return  {
+        who: AccountId32,
+        value: sts.bigint(),
+        isSettled: sts.boolean(),
+    }
+})
+
+export const MarketDisputeMechanism: sts.Type<MarketDisputeMechanism> = sts.closedEnum(() => {
+    return  {
+        Authorized: sts.unit(),
+        Court: sts.unit(),
+    }
+})
+
+export const OutcomeReport: sts.Type<OutcomeReport> = sts.closedEnum(() => {
+    return  {
+        Categorical: sts.number(),
+        Scalar: sts.bigint(),
+    }
+})
+
+export const Report: sts.Type<Report> = sts.struct(() => {
+    return  {
+        at: sts.bigint(),
+        by: AccountId32,
+        outcome: OutcomeReport,
+    }
+})
+
+export const MarketStatus: sts.Type<MarketStatus> = sts.closedEnum(() => {
+    return  {
+        Active: sts.unit(),
+        Closed: sts.unit(),
+        Disputed: sts.unit(),
+        Proposed: sts.unit(),
+        Reported: sts.unit(),
+        Resolved: sts.unit(),
+    }
+})
+
+export const ScoringRule: sts.Type<ScoringRule> = sts.closedEnum(() => {
+    return  {
+        AmmCdaHybrid: sts.unit(),
+        Parimutuel: sts.unit(),
+    }
+})
+
+export const Deadlines: sts.Type<Deadlines> = sts.struct(() => {
+    return  {
+        gracePeriod: sts.bigint(),
+        oracleDuration: sts.bigint(),
+        disputeDuration: sts.bigint(),
+    }
+})
+
+export const MarketPeriod: sts.Type<MarketPeriod> = sts.closedEnum(() => {
+    return  {
+        Block: Range,
+        Timestamp: Range,
+    }
+})
+
+export const Range: sts.Type<Range> = sts.struct(() => {
+    return  {
+        start: sts.bigint(),
+        end: sts.bigint(),
+    }
+})
+
+export const MarketType: sts.Type<MarketType> = sts.closedEnum(() => {
+    return  {
+        Categorical: sts.number(),
+        Scalar: RangeInclusive,
+    }
+})
+
+export const RangeInclusive: sts.Type<RangeInclusive> = sts.struct(() => {
+    return  {
+        start: sts.bigint(),
+        end: sts.bigint(),
+    }
+})
+
+export const Perbill = sts.number()
+
+export const MarketCreation: sts.Type<MarketCreation> = sts.closedEnum(() => {
+    return  {
+        Advised: sts.unit(),
+        Permissionless: sts.unit(),
+    }
+})
+
+export const SwapEvent: sts.Type<SwapEvent> = sts.struct(() => {
+    return  {
+        assetAmountIn: sts.bigint(),
+        assetAmountOut: sts.bigint(),
+        assetBound: sts.option(() => sts.bigint()),
+        assetIn: Asset,
+        assetOut: Asset,
+        cpep: CommonPoolEventParams,
+        maxPrice: sts.option(() => sts.bigint()),
+    }
+})
+
+export interface SwapEvent {
+    assetAmountIn: bigint
+    assetAmountOut: bigint
+    assetBound?: (bigint | undefined)
+    assetIn: Asset
+    assetOut: Asset
+    cpep: CommonPoolEventParams
+    maxPrice?: (bigint | undefined)
+}
+
+export interface CommonPoolEventParams {
+    poolId: bigint
+    who: AccountId32
+}
+
+export const PoolAssetEvent: sts.Type<PoolAssetEvent> = sts.struct(() => {
+    return  {
+        asset: Asset,
+        bound: sts.bigint(),
+        cpep: CommonPoolEventParams,
+        transferred: sts.bigint(),
+        poolAmount: sts.bigint(),
+    }
+})
+
+export interface PoolAssetEvent {
+    asset: Asset
+    bound: bigint
+    cpep: CommonPoolEventParams
+    transferred: bigint
+    poolAmount: bigint
+}
+
+export const PoolAssetsEvent: sts.Type<PoolAssetsEvent> = sts.struct(() => {
+    return  {
+        assets: sts.array(() => Asset),
+        bounds: sts.array(() => sts.bigint()),
+        cpep: CommonPoolEventParams,
+        transferred: sts.array(() => sts.bigint()),
+        poolAmount: sts.bigint(),
+    }
+})
+
+export interface PoolAssetsEvent {
+    assets: Asset[]
+    bounds: bigint[]
+    cpep: CommonPoolEventParams
+    transferred: bigint[]
+    poolAmount: bigint
+}
+
+export const Pool: sts.Type<Pool> = sts.struct(() => {
+    return  {
+        assets: sts.array(() => Asset),
+        status: PoolStatus,
+        swapFee: sts.bigint(),
+        totalWeight: sts.bigint(),
+        weights: sts.array(() => sts.tuple(() => [Asset, sts.bigint()])),
+    }
+})
+
+export const PoolStatus: sts.Type<PoolStatus> = sts.closedEnum(() => {
+    return  {
+        Closed: sts.unit(),
+        Open: sts.unit(),
+    }
+})
+
+export type PoolStatus = PoolStatus_Closed | PoolStatus_Open
+
+export interface PoolStatus_Closed {
+    __kind: 'Closed'
+}
+
+export interface PoolStatus_Open {
+    __kind: 'Open'
+}
+
+export interface Pool {
+    assets: Asset[]
+    status: PoolStatus
+    swapFee: bigint
+    totalWeight: bigint
+    weights: [Asset, bigint][]
+}
+
+export const CommonPoolEventParams: sts.Type<CommonPoolEventParams> = sts.struct(() => {
+    return  {
+        poolId: sts.bigint(),
+        who: AccountId32,
+    }
+})
+
+export const AccountId32 = sts.bytes()
+
+export const Asset: sts.Type<Asset> = sts.closedEnum(() => {
+    return  {
+        CategoricalOutcome: sts.tuple(() => [sts.bigint(), sts.number()]),
+        CombinatorialOutcomeLegacy: sts.unit(),
+        CombinatorialToken: sts.bytes(),
+        ForeignAsset: sts.number(),
+        ParimutuelShare: sts.tuple(() => [sts.bigint(), sts.number()]),
+        PoolShare: sts.bigint(),
+        ScalarOutcome: sts.tuple(() => [sts.bigint(), ScalarPosition]),
+        Ztg: sts.unit(),
+    }
+})
+
+export const ScalarPosition: sts.Type<ScalarPosition> = sts.closedEnum(() => {
+    return  {
+        Long: sts.unit(),
+        Short: sts.unit(),
+    }
+})
